Get Groups By User ID
We're currently getting groups by the group's id itself, not by the owner of the groups. We could make a different endpoint to get a specific group by it's ID if we need it, but I don't think we will. We'll fetch them as soon as they log in and keep them in state from dashboard, so we should never need it.